realtime-media-processor/backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
opencv-python-headless
tensorflow
